Estebanreyl/dev merge fixes  (#43)
* Merge fixes to acquire latest telemetry items

* Updated to master AzureUtilityManager
